dashboard-service-guyujung api 추가

+++ dashboard-service/src/main/java/com/example/vo/Contribution.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class Contribution {
     private Long id;
-    private Long contribution;
+    private float contribution;
 }
